double

+++ src/main/java/com/rk/rkstuff/accelerator/tile/TileAcceleratorCaseFluidIO.java
@@ -80,7 +80,7 @@ public class TileAcceleratorCaseFluidIO extends TileRK implements IMultiBlockMas
     }
 
     @Override
-    public int receiveCoolant(ForgeDirection from, int maxAmount, float temperature, boolean simulate) {
+    public int receiveCoolant(ForgeDirection from, int maxAmount, double temperature, boolean simulate) {
         if (!hasMaster() || isOutput) return 0;
         return master.receiveCoolant(side, maxAmount, temperature, simulate);
     }
