+++ src/main/java/com/rk/rkstuff/coolant/tile/ICoolantReceiver.java
@@ -4,6 +4,6 @@ import net.minecraftforge.common.util.ForgeDirection;
 
 public interface ICoolantReceiver extends ICoolantConnection {
 
-    public int receiveCoolant(ForgeDirection from, int maxAmount, float temperature, boolean simulate);
+    public int receiveCoolant(ForgeDirection from, int maxAmount, double temperature, boolean simulate);
 
 }
