+++ src/main/java/com/rk/rkstuff/solar/tile/TileSolarInput.java
@@ -43,7 +43,7 @@ public class TileSolarInput extends TileRK implements ICoolantReceiver, IMultiBl
     }
 
     @Override
-    public int receiveCoolant(ForgeDirection from, int maxAmount, float temperature, boolean simulate) {
+    public int receiveCoolant(ForgeDirection from, int maxAmount, double temperature, boolean simulate) {
         if (master == null) return 0;
         return master.receiveCoolant(from, maxAmount, temperature, simulate);
     }
